+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Release Versions:
 - Add forward force parameter to Impedance controller (#36)
 - Simplify and consolidate ostream operators (#37)
 - Bind inverse for each Cartesian class individually (#52)
+- Refactor get/set variable helpers in Cartesian state (#39)
 
 ## 6.3.1